+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Set `extensionKind` to `workspace` in `package.json`.
 - Update `README.md` to include bundled dependencies url.
 - Remove `vue` from `disableLanguages` because Vue is supported in Prettier 1.15 version.
+- Move `eslint`, `tslint` and `stylelint` integrations to separate file `integrations.ts` (out from `utils.ts`) and `require()` (load) dependency only when it's needed and cache it in-memory.
 
 ## 2.0.0 (August 16, 2019)